+++ src/creature.rs
@@ -11,11 +11,10 @@ use crate::metric::system::*;
 use crate::round::RoundState;
 use crate::space::{Position, PositionClaim};
 use crate::team::{EntityAddition, TeamId, TeamRules};
-use crate::util::Id;
+use crate::util::{collect_from_iter, Id};
 #[cfg(feature = "serialization")]
 use serde::{Deserialize, Serialize};
 use std::any::Any;
-use std::collections::hash_map::Values;
 use std::collections::HashMap;
 use std::fmt::{Debug, Formatter, Result};
 
@@ -45,16 +44,6 @@ pub struct Creature<R: BattleRules> {
 }
 
 impl<R: BattleRules> Creature<R> {
-    /// Returns an iterator over abilities.
-    pub fn abilities(&self) -> Values<AbilityId<R>, Ability<R>> {
-        self.abilities.values()
-    }
-
-    /// Returns the ability with the given id.
-    pub fn ability(&self, id: &AbilityId<R>) -> Option<&Ability<R>> {
-        self.abilities.get(id)
-    }
-
     pub(crate) fn set_team_id(&mut self, id: TeamId<R>) {
         self.team_id = id;
     }
@@ -64,8 +53,11 @@ impl<R: BattleRules> Id for Creature<R> {
     type Id = CreatureId<R>;
 
     fn id(&self) -> &CreatureId<R> {
-        let EntityId::Creature(id) = &self.id;
-        id
+        if let EntityId::Creature(id) = &self.id {
+            id
+        } else {
+            panic!("constraint violated: creature's id has a wrong type")
+        }
     }
 }
 
@@ -228,24 +220,6 @@ impl<R: BattleRules> Clone for CreateCreature<R> {
 }
 
 impl<R: BattleRules> CreateCreature<R> {
-    /// Collects an iterator into an hashmap.
-    /// Subsequent values with same key are ignored.
-    fn collect_from_iter<I>(
-        it: I,
-    ) -> HashMap<<<I as Iterator>::Item as Id>::Id, <I as Iterator>::Item>
-    where
-        I: Iterator,
-        <I as Iterator>::Item: Id,
-    {
-        let mut map = HashMap::new();
-        for e in it {
-            if !map.contains_key(e.id()) {
-                map.insert(e.id().clone(), e);
-            }
-        }
-        map
-    }
-
     /// Returns a trigger for this event.
     pub fn trigger<'a, P: EventProcessor<R>>(
         processor: &'a mut P,
@@ -324,14 +298,14 @@ impl<R: BattleRules + 'static> Event<R> for CreateCreature<R> {
             &mut battle.entropy,
             &mut battle.metrics.write_handle(),
         );
-        let statistics = CreateCreature::<R>::collect_from_iter(it);
+        let statistics = collect_from_iter(it);
         // Abilities' generation is influenced by the given abilities_seed, if present.
         let it = battle.rules.actor_rules().generate_abilities(
             &self.abilities_seed,
             &mut battle.entropy,
             &mut battle.metrics.write_handle(),
         );
-        let abilities = CreateCreature::<R>::collect_from_iter(it);
+        let abilities = collect_from_iter(it);
         // Create the creature.
         let creature = Creature {
             id: EntityId::Creature(self.id.clone()),
@@ -352,7 +326,7 @@ impl<R: BattleRules + 'static> Event<R> for CreateCreature<R> {
             &mut battle.entropy,
             &mut battle.metrics.write_handle(),
         );
-        // Add the creature to the actors.
+        // Add the creature to the entities.
         battle
             .state
             .entities
@@ -804,6 +778,7 @@ mod tests {
 
     impl<R: BattleRules> CharacterRules<R> for CustomCharacterRules {
         type CreatureId = u32;
+        type ObjectId = ();
         type Statistic = SimpleStatistic<u32, u32>;
         type StatisticsSeed = ();
         type StatisticsAlteration = ();